refactor: modify base package
