+++ module-api/build.gradle
@@ -19,9 +19,6 @@ jib {
     }
 }
 
-group = 'com.codingbottle'
-version = '0.0.1-SNAPSHOT'
-
 repositories {
     mavenCentral()
 }
