--- module-api/src/main/java/com/codingbottle/api/TestPing.java
@@ -1,12 +1,0 @@
-package com.codingbottle.api;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TestPing {
-    @GetMapping("/ping")
-    public String ping() {
-        return "pong";
-    }
-}
