+++ module-api/src/main/java/com/codingbottle/config/FirebaseInitializer.java
@@ -1,4 +1,4 @@
-package com.codingbottle.api.config;
+package com.codingbottle.config;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
