+++ module-api/src/main/java/com/codingbottle/domain/category/controller/CategoryController.java
@@ -1,6 +1,6 @@
-package com.codingbottle.api.domain.category.controller;
+package com.codingbottle.domain.category.controller;
 
-import com.codingbottle.api.domain.category.service.CategoryService;
+import com.codingbottle.domain.category.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
