+++ module-core/build.gradle
@@ -2,9 +2,6 @@ plugins {
     id 'java'
 }
 
-group = 'com.codingbottle'
-version = '0.0.1-SNAPSHOT'
-
 repositories {
     mavenCentral()
 }
