+++ module-core/src/main/java/com/codingbottle/auth/domain/entity/Role.java
@@ -1,0 +1,5 @@
+package com.codingbottle.auth.domain.entity;
+
+public enum Role {
+    ROLE_ADMIN, ROLE_USER
+}
