+++ module-core/src/main/java/com/codingbottle/auth/domain/entity/User.java
@@ -1,23 +1,20 @@
-package com.codingbottle.core.auth.domain.entity;
+package com.codingbottle.auth.domain.entity;
 
 import com.google.firebase.auth.FirebaseToken;
 import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.BatchSize;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.*;
-import java.util.stream.Collector;
-import java.util.stream.Collectors;
 
 @Getter
 @Entity
-@Table(name = "user")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Builder
 public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -79,5 +76,14 @@ public class User implements UserDetails {
         this.email = token.getEmail();
         this.name = token.getName();
         this.picture = token.getPicture();
+    }
+
+    @Builder
+    public User(String username, Role role, String email, String name, String picture) {
+        this.username = username;
+        this.role = role;
+        this.email = email;
+        this.name = name;
+        this.picture = picture;
     }
 }
