+++ module-core/src/main/java/com/codingbottle/auth/domain/repository/UserRepository.java
@@ -1,11 +1,9 @@
-package com.codingbottle.core.auth.domain.repository;
+package com.codingbottle.auth.domain.repository;
 
-import com.codingbottle.core.auth.domain.entity.User;
+import com.codingbottle.auth.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
-@Repository
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
 }
