--- module-core/src/main/java/com/codingbottle/core/HelfyCoreApplication.java
@@ -1,11 +1,0 @@
-package com.codingbottle.core;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HelfyCoreApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(HelfyCoreApplication.class, args);
-    }
-}
