--- module-core/src/main/java/com/codingbottle/core/auth/domain/entity/Role.java
@@ -1,5 +1,0 @@
-package com.codingbottle.core.auth.domain.entity;
-
-public enum Role {
-    ROLE_ADMIN, ROLE_USER
-}
